core: project structure
1. Removed the redundant ".pyc" files together with the
"__pycache__" folder.
Those are Python modules compiled into bytecode and
cached for optimization.
I doubt they need to be part of the repository, especially that they
were compiled with an old (and already unsupported) Python 3.8 version
vs. the latest Python 3.13 version.
2. Added a ".gitignore" file with a list of files and folders
to be ignored by "git" at the project level.

Signed-off-by: Pavel Bar <pbar@redhat.com>