refactor: code improvements & warnings fixes
1. Refactored the "_read_video_frame()" method
to be static to fix a warning, see [2].
2. Using the class name to access class-level constants.
3. Using module level "dunders" to state the author
and additional module information. This is probably a more
elegant solution compared to using comments and/or
a dedicated "creator()" method (the latter was removed).
More info at the "PEP 8 – Style Guide for Python Code", see [1].
4. Fixed a few additional warnings, see [3], [4] & [5].
5. Minor "README.md" update.

[1] https://peps.python.org/pep-0008/#module-level-dunder-names
[2] Method '_read_video_frame' may be 'static'
[3] Ui_MainWindow::setupUi()
Shadows name 'MainWindow' from outer scope
[4] AnalogGaugeWidget::set_MinValue()
Shadows built-in name 'min'
[5] AnalogGaugeWidget::set_MaxValue()
Shadows built-in name 'max'

Signed-off-by: Pavel Bar <pbar@redhat.com>